fix: make cc/uaa endpoints configurable
Make the cloud controller & uaa endpoints (required for ssh-proxy)
configurable, and default to hostnames as (being separte helm charts and
not a cf-operator/quarks generated pod)  we do not have access to a DNS
server which can resolve *.cf.internal addresses.

Also update the default doppler endpoint, for the same reason.